+++ CHANGELOG.md
@@ -4,6 +4,7 @@ https://keepachangelog.com/en/1.0.0/
 ## [1.33.0] - 2026-03-14
 - `StructWithoutElementaryMethodsOverridden`: Turned off by default
 - `DivideIntegerByInteger`: Set to `Info` by default
+- `HttpClientInstantiatedDirectly`: No longer triggers in test classes (xUnit, NUnit, MSTest)
 
 ## [1.32.0] - 2026-03-14
 - `RedisResponseNotHandled`: Detect when Redis response objects are discarded without checking for errors